Moved CicularQueue to the DataStructures namespace
