+++ Bot/GameSense/IncomeTracker.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using Bot.DataStructures;
 using Bot.GameData;
 using Bot.UnitModules;
 using Bot.Utils;
